Symlinks script is updated to the latest version

+++ create-library-symlinks.sh
@@ -1,19 +1,25 @@
+set -e
+
 function symlink-library {
-  name=$1
+  if [ -z ${LIBRARIES_DIR+x} ]; then
+    echo "LIBRARIES_DIR must be set to the libraries directory path"
+  else
+    name=$1
 
-  echo
-  echo "Symlinking $name"
-  echo "- - -"
+    echo
+    echo "Symlinking $name"
+    echo "- - -"
 
-  for entry in ../libraries/raygun_client*; do
-    echo "- removing symlink: $entry"
-    rm $entry
-  done
+    for entry in $LIBRARIES_DIR/$name*; do
+      echo "- removing symlink: $entry"
+      rm $entry
+    done
 
-  echo "- creating symlinks"
-  ln -s $(PWD)/lib/raygun_client* $(PWD)/../libraries/
+    echo "- creating symlinks"
+    ln -s $(PWD)/lib/$name* $LIBRARIES_DIR/
 
-  echo "- - -"
-  echo "($name done)"
-  echo
+    echo "- - -"
+    echo "($name done)"
+    echo
+  fi
 }
